create an ahoy wrapper
