+++ src/utils/analytics.ts
@@ -1,40 +1,9 @@
 import {isFunction, isUndefined} from 'lodash'
-import getAccessTokenFromCookie from './getAccessTokenFromCookie'
+import configuredAhoy from './ahoy'
 
 export const track = (event: string, paramsOrCallback?, callback?) => {
-  const isServer = typeof window === 'undefined'
-  let ahoy
-
-  if (!isServer) {
-    const token = getAccessTokenFromCookie()
-    ahoy = window.ahoy = window.ahoy || {}
-    ahoy.configure({
-      urlPrefix: '',
-      visitsUrl: `/api/visits`,
-      eventsUrl: `/api/events`,
-      page: null,
-      platform: 'Web',
-      useBeacon: true,
-      startOnReady: true,
-      trackVisits: true,
-      cookies: true,
-      cookieDomain: process.env.NEXT_PUBLIC_DEPLOYMENT_URL,
-      headers: token
-        ? {
-            Authorization: `Bearer ${token}`,
-            'Ahoy-Visit': ahoy.getVisitId(),
-            'Ahoy-Visitor': ahoy.getVisitorId(),
-          }
-        : {
-            'Ahoy-Visit': ahoy.getVisitId(),
-            'Ahoy-Visitor': ahoy.getVisitorId(),
-          },
-      withCredentials: false,
-    })
-  }
-
-  return new Promise((resolve, reject) => {
-    if (isServer) reject()
+  return new Promise((resolve) => {
+    const ahoy = configuredAhoy()
     let wasCalled = false
 
     const params = isFunction(paramsOrCallback) ? {} : paramsOrCallback
